+++ docs/en/history.rst
@@ -6,6 +6,7 @@ dev
 
   - Incorporated patch to add ``-d`` option to
     :ref:`command-add2virtualenv`, contributed by :bbuser:`miracle2k`.
+  - Add ``-i`` option to :ref:`command-mkvirtualenv`.
 
 2.9